refactor STFPM loss
